requirements.txt
requests
kafka-python
python-dotenv
gunicorn
Flask
dropbox
firebase-admin